Rename HydrometerCorrectionWizard to HydrometerCorrection.
