+++ app/App.jsx
@@ -25,7 +25,7 @@
 import React from 'react';
 import SignIn from './components/SignIn.jsx';
 import GuestMenu from './components/GuestMenu.jsx';
-import HydrometerCorrectionWizard from './components/HydrometerCorrectionWizard.jsx';
+import HydrometerCorrectionWizard from './components/HydrometerCorrection.jsx';
 import PageHeader from './components/PageHeader.jsx';
 import PageFooter from './components/PageFooter.jsx';
 import PreferencesDialog from './components/PreferencesDialog.jsx';
